Added users

+++ routes.php
@@ -1,12 +1,16 @@
 <?php
 
-$router->get('/', 'HomeController@index');
-$router->get('/tasks/create', 'TaskController@create');
-$router->get('/tasks', 'TaskController@index');
-$router->get('/tasks/{id}', 'TaskController@show');
-$router->delete('/tasks/{id}', 'TaskController@delete');
-$router->put('/tasks/{id}', 'TaskController@update');
-$router->post('/tasks/create', 'TaskController@create');
+$router->get('/', 'HomeController@index', ['auth']);
+$router->get('/tasks/create', 'TaskController@create', ['auth']);
+$router->get('/tasks', 'TaskController@index', ['auth']);
+$router->get('/tasks/{id}', 'TaskController@show', ['auth']);
+$router->delete('/tasks/{id}', 'TaskController@delete', ['auth']);
+$router->put('/tasks/{id}', 'TaskController@update', ['auth']);
+$router->post('/tasks/create', 'TaskController@create', ['auth']);
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@register', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
+
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
